Pulled router actions into seperate module.
