+++ webofneeds/won-owner-webapp/src/main/webapp/app/actions/actions.js
@@ -69,6 +69,15 @@ import {
     addConstParams,
 } from '../configRouting';
 
+import {
+    stateBack,
+    stateGoAbs,
+    stateGoCurrent,
+    stateGoDefault,
+    stateGoKeepParams,
+    stateGoResetParams,
+} from './cstm-router-actions';
+
 // </utils>
 
 // <action-creators>
@@ -360,100 +369,5 @@ export function needsClose(needUri) {
             // go back to overview
             dispatch(actionCreators.router__stateGoResetParams('overviewPosts'))
         )
-    }
-}
-
-/**
- * Action-Creator that goes back in the browser history
- * without leaving the app.
- * @param dispatch
- * @param getState
- */
-function stateBack() {
-    return (dispatch, getState) => {
-        const hasPreviousState = !!getState().getIn(['router', 'prevState', 'name']);
-        if (hasPreviousState) {
-            history.back();
-        } else {
-            dispatch(actionCreators.router__stateGoResetParams('landingpage'));
-        }
-    }
-}
-
-/**
- * reset's all parameters but the one passed as arguments
- */
-function stateGoAbs(state, queryParams) {
-    return (dispatch, getState) => {
-        const currentParams = getState().getIn(['router', 'currentParams']);
-        dispatch(actionCreators.router__stateGo(
-            state,
-            addConstParams(queryParams, currentParams)
-        ))
-    }
-}
-
-/**
- * goes to new state and resets all parameters (except for "pervasive" ones like `privateId`)
- */
-function stateGoResetParams(state) {
-    return (dispatch, getState) => {
-        const currentParams = getState().getIn(['router', 'currentParams']);
-
-        console.log('routing from',
-            getState().getIn(['router', 'currentState', 'name']), currentParams,
-            ' to ', state, addConstParams(resetParams, currentParams)
-        );
-
-        dispatch(actionCreators.router__stateGo(
-            state,
-            addConstParams(resetParams, currentParams)
-        ))
-    }
-}
-
-function stateGoDefault() {
-    return (dispatch, getState) => {
-        if ( state.get('initialLoadFinished') ) {
-            if (state.getIn(['user', 'loggedIn'])) {
-                dispatch(actionCreators.router__stateGoResetParams('feed'));
-            } else {
-                dispatch(actionCreators.router__stateGoResetParams('landingpage'));
-            }
-        }
-    }
-}
-
-/**
- * goes to new state and keeps listed parameters at their current values
- */
-function stateGoKeepParams(state, queryParamsList) {
-    return (dispatch, getState) => {
-        const currentParams = getState().getIn(['router', 'currentParams']);
-        const params = Immutable.Map( // [[k,v]] -> Map
-            queryParamsList.map(
-                    p => [p, currentParams.get(p)] // get value per param
-            )
-        );
-        dispatch(actionCreators.router__stateGo(
-            state,
-            addConstParams(params, currentParams)
-        ))
-    }
-}
-
-/**
- * goes to current state, but changes the parameters
- * passed to this function.
- * @param queryParams
- */
-function stateGoCurrent(queryParams) {
-    return (dispatch, getState) => {
-        const currentState = getState().getIn(['router', 'currentState', 'name']);
-        const currentParams = getState().getIn(['router', 'currentParams']);
-        dispatch(actionCreators.router__stateGo(
-            currentState,
-            addConstParams(queryParams, currentParams)
-        ));
     }
 }
